Add files via upload
This was the final project from the python course I had to take as a prereq. To be honest this is all foreign and I have no clue what any of that does. I uploaded the most advanced code I still remember how to read. I completed the previous course almost a full year ago, I would have liked to go straight to this but with the physics and math I needed I prioritized taking those sequentially.